+++ src/azure-cli/azure/cli/command_modules/vm/commands.py
@@ -189,7 +189,7 @@ def load_command_table(self, _):
         g.wait_command('wait')
 
     with self.command_group('disk-encryption-set', compute_disk_encryption_set_sdk, client_factory=cf_disk_encryption_set, min_api='2019-07-01') as g:
-        g.custom_command('create', 'create_disk_encryption_set')
+        g.custom_command('create', 'create_disk_encryption_set', supports_no_wait=True)
         g.command('delete', 'delete')
         # g.command('update', 'update')
         g.show_command('show', 'get')